Remove `Bylines\Admin` in favor of consolidated `Bylines\Post_Editor`
This is similar to the pattern established in `Byline_Editor`
